refactor: separate renderVideo
renderVideo 클래스를 검색한 비디오 렌더링과 저장한 비디오 렌더링 두 부분으로 나눠서 각각 renderSearchVideo, renderSaveVideo 클래스로 분리했다. 새로 생성된 두 클래스는 renderVideo를 상속한다.
