+++ src/js/index.js
@@ -1,5 +1,11 @@
 import '../css/index.css';
-import RenderVideo from './renderVideo.js';
+import RenderSaveVideo from './renderSaveVideo';
+import RenderSearchVideo from './renderSearchVideo';
+import SaveVideo from './saveVideo';
+import SearchVideo from './searchVideo';
 
-// eslint-disable-next-line no-new
-new RenderVideo();
+const searchVideo = new SearchVideo();
+const saveVideo = new SaveVideo();
+
+new RenderSaveVideo(saveVideo);
+new RenderSearchVideo(searchVideo, saveVideo);
